Notify View and SongPlayer of Changes in SongProgression
Made thumbTrack to notify seekBar's View's onChange(which sets the
current Time of the song) whenever it changes(value changes), so that
onChange will update currentBuzzObject(setCurrentTime) to
currentTime(value). Also allowed scope to reference on-change function
by binding it to an inner scope of the directive, so that it can access
outer scope information.....but where is it even doing that? still
debugging. for some reason, volume is changing the playback position .