Still Conceptually and Programmatically Debugging
still confused as to why we need to observe changes of scope.value and
max...i'm guessing it's just for the fill style and thumStyle functions
to update the seekBar's ui.

My Guess:

     //View value attribute --> Model changes in value(currentTime)
                 //it's to UPDATE directive's ngStyle handlers about
chnages in currentTime and length of song(value and max specifically.)

//since the directive does not have access to changes in currentime
directly, but the view does? what???? that doesn't make sense. why must
it use view and max attributes. and why is that the attributes have
access to the Songplayer's values but directive's scope does not?